MS-81 feat(func): move function
